requirements.txt
matplotlib==3.6.3
pandas==1.5.2
psycopg2==2.9.5
psycopg2_binary==2.9.5
python-dotenv==0.21.0
setuptools==65.5.0
numpy==1.24.1
scipy==1.10.0